refactor, use pub+pub in demo
