+++ jason/e2e-demo/js/index.js
@@ -9,6 +9,15 @@ async function f() {
 
     caller_room_handle.on_new_connection(function (connection) {
         console.log("caller got new connection with member " + connection.member_id());
+        connection.on_remote_stream(function (stream) {
+            console.log("got video from remote member " + connection.member_id());
+
+            var video = document.createElement("video");
+
+            video.srcObject = stream.get_media_stream();
+            document.body.appendChild(video);
+            video.play();
+        });
     });
     caller.on_local_stream(function (stream, error) {
         if (stream) {
@@ -22,6 +31,17 @@ async function f() {
         }
     });
 
+    responder.on_local_stream(function (stream, error) {
+        if (stream) {
+            var video = document.createElement("video");
+
+            video.srcObject = stream.get_media_stream();
+            document.body.appendChild(video);
+            video.play();
+        } else {
+            console.log(error);
+        }
+    });
     responder_room_handler.on_new_connection(function (connection) {
         console.log("responder got new connection with member " + connection.member_id());
         connection.on_remote_stream(function (stream) {
