+++ jason/src/lib.rs
@@ -15,7 +15,7 @@ pub use console_error_panic_hook::set_once as set_panic_hook;
 
 #[doc(inline)]
 pub use self::{
-    jason::{Jason, RoomHandle, ConnectionHandle},
+    jason::{ConnectionHandle, Jason, RoomHandle},
     media::MediaStreamHandle,
 };
